% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlete_ranking_breakdown.R
\name{tri_athlete_ranking_breakdown}
\alias{tri_athlete_ranking_breakdown}
\title{This Athlete Ranking Breakdown method returns detailed ranking information for an athlete.
It provides the result of all events that are contained within a ranking criteria and contains
detailed information as to the status of that event.}
\usage{
tri_athlete_ranking_breakdown(rank_no, athlete_no)
}
\arguments{
\item{rank_no}{:ranking id}

\item{athlete_no}{:athlete number}
}
\value{
A data.frame
}
\description{
A basic athlete object and ranking listing object are returned with all rankings details
contained within the breakdown object.
}
\details{
The output of the breakdown object is ranking specific and you should consult each ranking
criteria for more information but the following guidelines may be used:

For events with multiple periods the current period is designated as 1 and the prior
period 2.
The event type corresponds to the points level listed on each ranking criteria which is
used to determine the number of scoring positions and maximum number of points.
For some ranking criteria the results of the event are modified according to the ranking
criteria (e.g. where non-continental athletes are in the results of a Continental event). In
these cases the prog_id refers to the results used and the original results are available in
the override_prog_id. Similarly the override_prog_time and original_position may be
used for the raw non-modified data.
The boolean include value displays whether this score has been included in the ranking
score
For events that are within a separate event e.g. a National Championship event taking
place in a Continental Cup the event_container key will show the event_id of the
containing event.
If the event has been modified and a deduction is applicable the % deduction will be
shown
The notes key contains specific information as to why the score has not been included if
applicable
}
\examples{
\dontrun{
  tri_athlete_ranking_breakdown(14, 7776)
}

}
\references{
\url{https://developers.triathlon.org/reference/athlete-ranking-breakdown}
}
