% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_listings.R
\name{tri_get_event_list}
\alias{tri_get_event_list}
\title{The Event Listings method returns a paginated, filterable list of basic event objects matching the search criteria.}
\usage{
tri_get_event_list(startdate, enddate)
}
\arguments{
\item{startdate}{date: start date to search for events}

\item{enddate}{date: end date to search for events}
}
\value{
A data.frame
}
\description{
The Event Listings method returns a paginated, filterable list of basic event objects matching the search criteria.
}
\examples{
\dontrun{
  tri_get_event_list('2023-01-01', '2023-02-01')
}

}
\references{
\url{https://developers.triathlon.org/reference/event-listings}
}
