% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_listings.R
\name{tri_program_list_get}
\alias{tri_program_list_get}
\title{The Program Listings method returns an array of basic program objects for an event id.}
\usage{
tri_program_list_get(no)
}
\arguments{
\item{no}{:event number}
}
\value{
A data.frame
}
\description{
The Program Listings method returns an array of basic program objects for an event id.
}
\examples{
\dontrun{
  tri_program_list_get(90162)
}

}
\references{
\url{https://developers.triathlon.org/reference/program-listings}
}
