% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_listing_detail.R
\name{tri_ranking_detail}
\alias{tri_ranking_detail}
\title{The Ranking Detail method returns the output of a specified ranking and contains all the
summary information required for basic ranking information. A Rankings Listing object is
returned containing the summary information about the ranking returned while the rankings
themselves are contained within a corresponding rankings object.}
\usage{
tri_ranking_detail(rank_id)
}
\arguments{
\item{no}{:ranking id}
}
\value{
A data.frame
}
\description{
The Ranking Detail method returns the output of a specified ranking and contains all the
summary information required for basic ranking information. A Rankings Listing object is
returned containing the summary information about the ranking returned while the rankings
themselves are contained within a corresponding rankings object.
}
\examples{
\dontrun{
  tri_ranking_detail(11)
}

}
\references{
\url{https://developers.triathlon.org/reference/ranking-detail}
}
