% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking_listing.R
\name{tri_ranking_list}
\alias{tri_ranking_list}
\title{The Rankings Listing method simply returns all available rankings and corresponding supplemental information such as the last published date.
You may restrict returned results by either category_id or region_id.
For weekly rankings an ISO week date is presented to identify the ranking week.}
\usage{
tri_ranking_list()
}
\value{
A data.frame
}
\description{
The Rankings Listing method simply returns all available rankings and corresponding supplemental information such as the last published date.
You may restrict returned results by either category_id or region_id.
For weekly rankings an ISO week date is presented to identify the ranking week.
}
\examples{
\dontrun{
  tri_ranking_list()
}

}
\references{
\url{https://developers.triathlon.org/reference/rankings-listing}
}
