% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{tri_results_list_get}
\alias{tri_results_list_get}
\title{The Results method searches a specified collection and returns a corresponding basic object for each.
We don’t recommend to merge the results of several queries in a single results list because their ranking criteria are often non-comparable: rather display them in separated sections.}
\usage{
tri_results_list_get(event_no, prog_no)
}
\arguments{
\item{event_no}{numeric: event number}

\item{prog_no}{numeric: program number}
}
\value{
A data.frame
}
\description{
The Results method searches a specified collection and returns a corresponding basic object for each.
We don’t recommend to merge the results of several queries in a single results list because their ranking criteria are often non-comparable: rather display them in separated sections.
}
\examples{
\dontrun{
  tri_results_list_get(90162, 270563)
}

}
\references{
\url{https://developers.triathlon.org/reference/search-results}
}
